feat(authentication): Add CRSF token view
Add "get_csrf_view" to protect POST, PUT, PATCH, DELETE views.
We opted for HTTP Only and Secure cookie authentication to store the cookie on the client side and avoid exposure to the XSS vulnerability that npm packages in the ReactJS application could offer.
To prevent Cross Site Forgery Requests, we opted for an anti-csrf Cookie configuration with setting.SAMESITE = "Strict".
And a whitelist for URLS that may request the anti CRSF token